backend/requirements.txt
fastapi==0.100.0
pydantic>=1.10fastapi==0.100.0       
uvicorn[standard]==0.23.1 
pydantic==1.10.9       
sqlalchemy==2.0.19     
python-dotenv==1.0.0           
python-jose==3.3.0
passlib[bcrypt]==1.7.4
